Minor doc cleanup, and add 'end' button back to test tour
'Forward reference' was referring to the direction in the JavaScript,
not to the UI; small clarification to another part too.

The difference between 'end' and leaving the automatic okay is that
clicking 'end' actually removes the tour from the cookie, whereas
an 'okay' button just hides it.

This makes a difference if you load the page again (without tour and
step parameters).

Change-Id: I078a14846155e98c1f77542f1e1fb6b7767623dd