FIx handleLinkClick to handle sub-elements within links
If the user clicks on an element descendant of the link, it will
bubble up, and hit the delegated listener on the <a> element.  However,
'target' will be set to what they actually clicked on.

This means accessing the href will generally yield undefined if they
don't click on the link.

However, currentTarget is always whatever element it has bubbled up to,
in this case the actual <a> element.

Bug: 71494
Change-Id: I317989871bddd3cc178bcc012988b1d66520dcba