Don't actually flip direction of page for directionality unit tests
This involves changing getShouldFlipHorizontally to take the two
directionalities as parameters.

The wrong values are still passed in when unrelated unit tests (that
just call getShouldFlipHorizontally in the course of things) run
(since the 'plain' QUnit page is not a MW page and is thus missing
standard attributes).

However, that doesn't affect anything, actual MW pages do the right
thing, and the tests work.

Change-Id: Ib73e60e52ba571c7f7fe6450d1412c964563d55b