Handle Help:Guided tour/guider not existing, not being able to launch
test tries to exercise all the main features, including the ability to
get the content of a tour step from a page.

However, not all wikis have this feature set up.  Rather than needing
to do a maintenance script to populate this and have wikis maintain it,
we can just skip this step if it's not present.

Similarly, the 'firstedit' tour only works on articles.  So if they're
not currently on an article, just don't demonstrate launching another
tour.

Bug: T153000
Change-Id: I089a481f7bc13263ebca8f50828b66a5d26e5c65