tests: Don't assume synchronous jQuery.Deferred
Most tests here already used the promise correctly, but this showStep()
tests didn't account for the fact that the spied target method is
called after initialize() inside showStep().

This fails under jQuery 3 because promises are enforced to be async.
Fix by moving the assertion to the same promise, and make sure
the test waits for it by using assert.async(), or by returning
the promise to QUnit.

Note: Step objects break if initialize() is called twice. Tour
objects try to account for this by tracking isInitialized once done
but if called a second time before the first call finished (which
would've happened here) then it will still call it a second time.

Work around by waiting for initialize() before calling showStep().
To be fixed in Tour.js proper.

Bug: T124742
Change-Id: I4e9b366613bc855bd63576f8a4637a462cffaf6a