Rename internal MakeGlobalVariablesScript hook callback
Rename it so that the callback itself shows up when searching
for MakeGlobalVariablesScript, and more generally for consistency
with other hook callbacks.

The prevous name made it appear as a generic method that could
be refactored when in-fact it had to fulfil a very specific contract
(although it could be refactored to its own method if that such
method is desirable).

Change-Id: I805b297d11cfb803b82f8a56208c2c2a13a5ffc6