Disable logging
I set shouldLog to false in the constructor, which seems to be
the only code path that builds a tour.  Just in case someone changes
that, I also set shouldLog to false in all the calls to the constructor,
so that someone refactoring or trying to re-enable would have a chance
to search for this change and undo it.

NOTE: if this instrumentation is ever updated, the migration to Event
Platform will probably have been completed, so it probably won't work.
We could just remove the schemas from extension.json and disable the
instrumentation more permanently.

Bug: T288416
Change-Id: I8c0adb16153205e5ce19aba10ae4dca5840dcacb