tests: Use assert.throws() directly with custom comparator
This ensures current and future improvements to native assert.throws()
are picked up, and also allows ESLint to recognise it as this kind
of assertion.

It also removes the amount of custom code to be maintained and trusted
for correctness of test results, and should make the test feel more
familiar to those less familiar with GuidedTour specifically.

Change-Id: I98207eb9a19ba716da65ae4c7db125f004a46912