Remove EventLoggingSchemas entries
The GuidedTour instrumentation was disabled in I8c0adb16153. At the time
of writing it is not clear when the instrumentation code itself will be
removed.

Fortunately, removing the EventLoggingSchemas entries is a change that
requires little time commitment from the reviewer and has a positive
performance impact - remembering that the EventLoggingSchemas entries
are still included in the ext.eventLogging RL module, which is fetched
on the majority of pageviews.

Bug: T303712
Change-Id: If129d916025cb1b629b52809aedce9211c5e3777