support dark-mode by using tokens for colors and cdx-button for close
This replaces all colors with the respective MediaWiki design tokens. In
cases where there was no perfect match, the next best token was chosen.

Also, this replaces the close-button with a quiet codex-button to make
sure the icon stays visible in dark mode too.

This leaves the `background-color: #000;` of the overlay unchanged,
because changing this to something bright in dark-mode makes things
worse to my eyes. It would brighten the entire screen, which is
undesirable here.

Also, this uses a token for the box-shadows. While those currently do
not change in dark-mode, that decision might be reconsidered in the
future, and so GuidedTours will automatically get the benefit if that
happens.

Bug: T370454
Change-Id: I33805963baa9c6f19728ea9bdede96d1b0295050